warning fixes

+++ src/include/input.hpp
@@ -3,6 +3,6 @@
 
 void inputThreadFunc(std::atomic_bool &stop);
 
-float get_paddle_movement();
+int get_paddle_movement();
 
 #endif //PONG_INPUT_H
